+++ Assets/Scripts/Interfaces/IDamagable.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    delegate void DamageDelegate(GameObject damager, float damage, Vector3 point, Vector3 direction);
+    delegate void DamageDelegate(GameObject damager, int damage, Vector3 point, Vector3 direction);
 
     event Action HealthChangeEvent;
     event DamageDelegate DamageEvent;
@@ -16,6 +16,6 @@ public interface IDamagable
     GameObject gameObject { get; }
     Transform transform { get; }
 
-    void Damage(GameObject damager, float damage, Vector3 point, Vector3 direction);
-    void Kill(GameObject killer, float damage, Vector3 point, Vector3 direction);
+    void Damage(GameObject damager, int damage, Vector3 point, Vector3 direction);
+    void Kill(GameObject killer, int damage, Vector3 point, Vector3 direction);
 }
